status: update dirstate mtime in Rust status
Summary:
Now the Rust status updates the dirstate mtime when clearing a tracked file's NEED_CHECK flag (for watchman case). This is a minor optimmization we didn't implement before. Having up-to-date metadata in the dirstate saves a disk comparison in some cases such as watchman fresh instance.

This required invalidating "now" mtimes in the treestate before saving to avoid race conditions. I also added support for the test fakedirstatewritetime extension which tests such as test-merge1.t use to force the race condition.

Reviewed By: quark-zju

Differential Revision: D46832705

fbshipit-source-id: b90c53511d7a47a1e05f9c3cabebe8ffed769abd