gotham_ext: move scuba population to earlier when possible
Summary:
Move population of scuba logging for items taken directly from the request to earlier in the processing (specifically, during `inbound` for scuba middleware).  This means these will be logged even if the request is cancelled.

Note that quite a few things are still missing that would be useful, as they are determined during or after request handling, but these will require more work to log them.

Reviewed By: clara-9

Differential Revision: D47511992

fbshipit-source-id: 61a39709eebb371d1d8ef582651451cbadd223e5