symlinks: make PrjFS handle symlink notifications properly
Summary:
When creating a new symlink manually on Windows, `PrjFS` used to categorize the newly created file as a regular file instead of a symlink, which this diff fixes.

Since symlink creation on Windows is racy, this diff double checks the type of file for regular files in case they are symlinks, in a similar way to what is done for directories.

Reviewed By: xavierd

Differential Revision: D46860405

fbshipit-source-id: 87aa8b3909553baabe8480f69e996682674c69e7