Add RepoBlobstoreUnlinkOps facet exposing unlink for rebo blobstores
Summary:
We already had a `RepoBlobstore` facet which exposed the `Blobstore` trait for the blobstore backing a repo.

Expand this concept to a new facet which exposes the `BlobstoreUnlinkOps` trait (`Blobstore` + `BlostorePutOps` + `BlobstoreUnlinkOps`)

This enables us to leverage unlinking on a repo blobstore through a consistent abstraction from other ways we have to interact with a blobstore.
The alternative would be to do all the low-level things exposed through this facet in-place, which would be mixing different levels of abstraction.

Reviewed By: RajivTS

Differential Revision: D47630536

fbshipit-source-id: 7be3bcd37f24335ae23d79b732830b4085c45edb