Only allow typing one diff number
Summary:
Add a "maxTokens" value to `type: 'field'` entries in the commit mesage schema. Then the UI won't render the text entry if there's already >= maxTokens values in there.

This is used mostly for the diff number, where we don't want to attach multiple. I think it will be useful for other fields too though, and making it support any number is just as easy as supporting one.

Reviewed By: quark-zju

Differential Revision: D47604220

fbshipit-source-id: c4fc012ce6a5c58e10fc35015a7a8dc32ed7dfa6