inodes: disallow concurrent checkout when resuming
Summary:
When EdenFS is interrupted mid-checkout either due to a crash, or simply via
`eden stop`, the information is recorded on disk to allow the next checkout to
be resumed. There is however a subtle race in this case if 2 FORCE checkout are
started at the same time. They both will detect that a previous checkout was
interrupted, and both will start a checkout operation. This is not the desired
behavior as one of the checkout should stop and warn that an existing checkout
is ongoing instead!

The fix is trivial: the checkoutPid should simply be updated when checkout is
starting.

Reviewed By: kmancini

Differential Revision: D46037833

fbshipit-source-id: a711a1575dcbc671185fde7d4114ee1633e53408