Fix white line when a commit is selected
Summary:
Design nit: if you select the topmost commit in a stack, it would render white (in light theme) in a line above the commit bubble. This is an attempt to cover up the darker line.

However, when selected, we're using the wrong color, so the white jumps out. Instead we need to use the selection bg color.

one complication is that the selection color is partially transparent, so we need to first apply a layer of bg white, then the selection color on that so that there's no problem with applying this on top of the selection bg layer.

Reviewed By: quark-zju

Differential Revision: D47643134

fbshipit-source-id: 51e0b5ab9b17ce7ce26ebded58d348dbd5d46c26