Add open file button to comparison view
Summary:
A couple of small changes to the comparison view

- add a button that opens the file. (we had this on each line, but if you're not viewing a comparison against head, those line numbers don't line up... so we disallow per line file open but add a button at the header for opening the file in general)
- change the title text to an actual tooltip for copying the path. The title text took too long to appear, the tooltip is a much smoother experience
- shorten commit hash in comparison name

Note that this whole component is made with the Primer design system. Eventually I think we should scrap this and rebuild it with our new comparison view being used for split/partial commit. It's really weird to have a totally different design system that we have to load in.
We're not even sharing with reviewstack.dev, there's really no point in using it in this way.

Reviewed By: quark-zju

Differential Revision: D47600367

fbshipit-source-id: e0335cab2c150370c0663e32950a600ed2cda5d9