nfs: fix use after free in takeoverStop
Summary:
Looking at the root cause of recent crashes, one of the main one appears to be
a use after free in RpcServer::takeoverStop. The local root cause of which is
caused by the use of a `this` capture, moving code around makes the use after
free disappear.

Taking a step back and trying to understand what the actual root cause of the
bug is is somewhat tricky, but my suspicion has to do with
`EdenMount::fsChanelInitSuccessful` where the `FsChannel` is being destroyed
manually. In the case where `Nfsd3::destroy` enqueues the `delete this` before
the `takeoverStop` enqueues the lambda we could end up in a situation where
Nfsd3 is freed (and thus the RpcServer destroyed) before `takeoverStop`
completes and thus causing this bug.

Reviewed By: kmancini

Differential Revision: D46912289

fbshipit-source-id: f0fd1c6daf897d5fdd8169bc1cfd9fb667d97afd