Define In-memory InodeCatalog
Summary:
Adding an in-memory `InodeCatalog`. This will be used to see if we can succuessfully speed up various scm operations on Windows (e.g., `hg update` and `hg rebase`). It might also be useful for Sandcastle builds where persistence is not the aim.

This first commit introduces this new class - `MemInodeCatalog`. Subsequent commits will add implementation, tests, and integration with `Overlay`.

Reviewed By: xavierd

Differential Revision: D47601736

fbshipit-source-id: 751c79b7d24fc6dae1253ed3cb621e5fc020568f