reviewstack: fix github stacked diff display
github diffs were showing the full set of commits in the stack rather than just the commits for the given PR version

gitHubPullRequestComparableVersions was returning null for base commit,  which meant we always fell back on the } else if (afterBaseCommitID != null) { branch of gitHubPullRequestVersionDiff.  That fallback used to work fine, but now its giving us the base of the full stack rather than the base of the PR. I don't know how it worked before (perhaps github api behaviour changed?), but there you go!

Fix is to use the version.baseParent in gitHubPullRequestComparableVersions, which has the right value in it.

Test Plan:

yarn test --watchAll=false

Local build and run, I'll add some pictures on the PR