Get latest commit message from selector
Summary:
Context for this stack:

This is the beginning of a stack working on message syncing. The idea is that we'll use the remote (phabricator/github) diff message as the source of truth of the commit message for a diff. (this technique has worked for us in the past and I'd like to continue using it)

This means when we render a commit message locally, we'll first lookup the message on phabricator/github, and if there's a message, use that instead of the local commit message. When we submit a diff, we use the local commit message as the diff/PR message anyway. So by default, the two are in sync.

From there, they may diverge in two ways:
(1) you edit the remote message
(2) you edit the local commit message

The approach of using the remote message as a source of truth means we support (1) by default.

For (2), if we assume you always edit your commit message through ISL (!), then we can easily support this by making teh starting point of your edit be the remote message. Then on save/confirm, your local edits can be pushed up to the remote diff/PR.

This push can be done with a mutation (internally jf, for github with the github graphql API). Or, if you're submitting, then by adding the "update fields" flag for the jf submit.

This then only leaves the case where you edit the commit message locally outside of ISL. I believe we don't need to hyper focus on this case, but we should have a reasonable way of interacting with this situation.

To handle this, we'll show a little banner in the commit info view when the local message differs from the remote one. This banner will have a "load local message" button, which will set all the fields to be edited and load the local contents. Then, you can hit save / amend message to update the remote message to match (which you wanted to do anyway, right?)

Some cases worth noting:
- optimistic state is another confounding factor. Generally, we actually want to prefer optimistic state over local OR remote messages (in the event that the optimistic message is different than the local one). For example, if you run "amend", it will optimistically show the new message. But if you had modified it remotely previously, that optimistic state would get hidden. Our assumption that we start all edits from the remote message means that we should expect the message in the "amend" to always be the latest. Thus, we prefer the optimistic message. Eventually, the new message would be pushed to the remote and then next fetch will then agree with the optimistic message.
- we'll need to be careful of any command that incidentally affects commit messages. For example, "edit stack" will need ot track which commit messages it updates for commits with diffs attached, then run the external mutation to update those messages.
- linking and unlinking diffs constitutes a decision about whether to keep using the remote message. This will require some special care. Notably, if you amend commit message to include a new diff number, we probably need to immediately update the remote message with your new local message.

----------

This diff in particular begins the ground work for showing remote messages locally. Instead of using the commit message from CommitInfo.description, we'll add a recoil selector that can map commit hash -> remote message or otherwise local message.

Reviewed By: sggutier

Differential Revision: D47453562

fbshipit-source-id: 702b4fa7319620103f92c98e0fd43d753addb728