Add context menu to load original commit message
Summary:
If we're showing the newer remote commit message, we need a way to load the local message instead.

We could either (1) have this change the rendering to show the original message, or (2) load the contents into the commit fields in edit mode.

(2) is a slightly simpler UX, I think. It lets you see your local message, then confirm its what you want by amending (metaedit). This reduces the need for a special toggle mode for whether you're seeing the remote message.

Reviewed By: quark-zju

Differential Revision: D47526408

fbshipit-source-id: 523c95ce2777f11e4420c5cb48c0d913bb698760