Disable message syncing based on code review provider
Summary:
For GitHub, we don't want to do message syncing. Internally, we do.

Techincally, GitHub could support message syncing, but it's a problem with the mental model.

Specifically, GitHub does not sync the PR message onto the commits before landing. This means that it does not really make sense to have GitHub as the source of truth for _commit_ messages, since there's no guarantee the two will end up in sync upon landing.

We could arguably do something like this syncing on land in `sl ghstack land`.... but we're not doing that right now so let's not bother with syncing.

We could also make syncing optional or something, since the syncing workflow is probably useful (showing remote changes locally). But for now let's go simple.

We'll have the codereviewprovider define whether it enables or disables message syncing.

Reviewed By: quark-zju

Differential Revision: D47562203

fbshipit-source-id: fcf1714fdecab49023172a91d39b1bceeb70d8a2