Update remote commit message on metaedit
Summary: If you do an "amend message" (metaedit) on a commit with a diff attached, we need to update the diff's remote message as well in order to keep it in sync. Otherwise, the remote message would immediately be used locally instead of your newly typed message, making it look like you didn't do anything.

Reviewed By: quark-zju

Differential Revision: D47610766

fbshipit-source-id: 8f02409e776d4e9f1b37a03795109673d4a727d2