Sync message when doing amend&submit
Summary:
Continuing with message syncing, we need to sync the local message to the remote whenever we modify the commit message. Notably:
- amend
- amend & submit
- amend message (metaedit)
  - related but needs extra handling: attaching a new diff number to a commit via metaedit

These cases do not require syncing:
- submit without amending the local commit
- committing
- committing and submitting

Technically, we could handle all the cases by mutating the remote message during any amend operation. The small extra change we do in this diff is to take advantage of "amend & submit", and just have the submit do the sync for us (since jf submit knows how to update fields). This is simpler and works as a single submit transaction and is more similar to how most users would submit their diffs.

By doing this through the submit operation, we'll only need to mutate directly when the users runs amend or metaedit without submitting.

Note that I don't bother implementing the `--update-fields` flag for GitHub submits, since we never use them.

Reviewed By: quark-zju

Differential Revision: D47568745

fbshipit-source-id: 2095ade69149325a7b27f118507257d1d5bb1390