Banner for showing different remote message
Summary:
We now show the remote message if it's newer. If we do this, we should let the user know that this message is different from the local message, and give them a chance to load the old message in case it was what they wanted.

Here I add a banner at the top of the commit info view if the message is substantialy different from the local version (and therefore being loaded from remote)

Note that we pass in the commit message here instead of looking it up by hash. This means we fetch the commit with previews, as intended. So if a commit has a preview that changes the commit message, it won't think that it has new remote changes.

Reviewed By: quark-zju

Differential Revision: D47525699

fbshipit-source-id: 04ad8bf4c401a3c982042eb6d91fb5219c53fd4e