Button to amend message with remote summary
Summary:
In the previous diff, we added a way to load the local message instead of the remote one. We can also offer the opposite: a way to confirm, "yes I'd like the remote message", and amend that message locally.

Doing this runs metaedit with the new message. In effect, this removes the "Showing remote message" banner, since now the local message agrees with the remote one.

Reviewed By: quark-zju

Differential Revision: D47556779

fbshipit-source-id: fc0cf191677feebe44176f2f89cd900209405ef9