Use diff's latest title for commit title
Summary:
In the preivous diff, we started using the remote message in the commit info. But the commit title should also use the new value in the main smartlog commit list.

Here I add a new recoil selector for just titles, which workes similarly to descriptions. Then I use this in the commit list.

Reviewed By: quark-zju

Differential Revision: D47477969

fbshipit-source-id: 1d7e0425c97573f196625bbd3d835ad710ae03c4