Fetch indivual diff summary when viewing commit info view
Summary:
Small UX improvement, we should refresh the currently visible commit's diff summary when you open it in the sidebar. This way, even if we do the main fetch for ALL diff infos on a slowish timer, we'll show the latest info for any commit you click on. This is most useful for the commit message, which may have changed remotely. But also technically applies to everything else in the fetch (diff status, comment count, etc)

Previously we threw out all diff summaries when we got new data. Now, we can mix&match partial diff fetches. So we need to keep our old data too and only rewrite the newly fetched entry.

Reviewed By: quark-zju

Differential Revision: D47475698

fbshipit-source-id: 5734c4b86838b4173a0d8b81bb45df2f96697e2f