Add query for phabricator's commit message
Summary:
See context in D47453562

We'll need to query for the remote commit message in order to show it instead of the local one.

Here I add the commit message to our existing diff summary fetch. This does increase the size of the graphql response by a bit, since we're fetching all diff summaries now which are longish text blobs. But still it's not all that much data being fetched, so I think it's ok.

The alternative approach would be to only fetch the commit messages as part of a separate data fetch, which we store separately. Then we only fetch for one diff at a time when viewing it in the commit info view.

Reviewed By: quark-zju

Differential Revision: D47453734

fbshipit-source-id: 49da130ee15f763bd4184ce9e76b5f01c7fe011a