stop constructing blobs with their IDs
Summary:
It is a minor historical mistake that Blob knew its own ID. There can
be multiple ID schemes that reference a single blob, and sometimes we
want to construct a blob without an ID available.

Now that blob IDs are never read directly, remove the field and
simplify a variety of construction sites.

Reviewed By: kmancini

Differential Revision: D39188561

fbshipit-source-id: 72b59b744aac42c312816d568fa563629575267a