Handle the case of empty Committer in Mononoke
Summary: In the current implementation, if the committer in the Mononoke commit was empty, we reused the author as the committer in the Git commit. This breaks the mapping between the Mononoke and the Git commit since if the Git commit was translated back to Mononoke, it would have both author and committer fields which is invalid. This diff makes sure that if the committer is absent in the Mononoke commit, then we use the empty Git signature for the committer in the Git commit as well.

Reviewed By: mitrandir77

Differential Revision: D47716705

fbshipit-source-id: 76889b7a8b0c286f7162aeb669c93a9562188d02