Introduce type for representing Git commit
Summary: Now that we have a derived data type for Git Trees that stores the raw git tree based on a `BonsaiChangeset`, this diff introduces a type that would be used to represent Git commit for a given `BonsaiChangeset`. The type by itself would just be changeset-based pointer to the raw Git commit that we would generate and store. Converting this type into derived data and implementing the generation logic will come in the next diff.

Reviewed By: mitrandir77

Differential Revision: D47397270

fbshipit-source-id: 7aee0d3ab32430ac904d7700f367ccfb58ebcf7e