Implement core commit derivation logic
Summary: Now that the building blocks are in place, this diff implements the core Git commit derivation logic. It relies on `bonsai_git_mapping` to determine if the mapped Git commit already exists or not and every new derived commit is stored through the same interface. The actual commit derivation involves creating the `git_object::Commit` object from the available `BonsaiChangeset` and the `GitTree` corresponding to this commit. This logic is missing unit tests which will be added in the next diff.

Differential Revision: D47515134

fbshipit-source-id: 23a84995107235666a4c5800f724c67e5964a8f7