Make MappedCommitId derivable
Summary: There is a whole lot of plumbing that goes into making a type derivable. I did not want to club together this plumbing logic with actual derivation logic which is why this diff just contains the derivation skeleton for `MappedCommitId`. The core derivation logic will be introduced in the next diff. This type will be disabled by default so its ok that the derivation logic currently has a `todo!` in it.

Reviewed By: mitrandir77

Differential Revision: D47400814

fbshipit-source-id: f39ecc441b25df32033af05070f66ba775d25e0c