rework deleted manifest unittest
Summary: This test calls low-level derivation primitives that don't operate on commit but rather on trees of all changes between commit and it's parents. I want to change the API so let's change the tests to operate on higher level and don't depend on internal details so much. I'm changing the tested commits a bit so the hashes do change.

Reviewed By: RajivTS

Differential Revision: D47646716

fbshipit-source-id: 7107e6fe3dd7884e05142444b3bed24fdfda9ffa