Store recently used reviewers
Summary:
Add storage for recently used reviewers, so that we can show them. Anytime you use a reviewer it will get saved to recent, then persisted to local storage. We then sort by number of uses and show them as suggestions.

Note: localstorage may not be available all the time, so we try/catch. I don't love the idea of storing serialized JSON to user-visible `sl config`... so I went with a more temporary localstorage. Perhaps we need to think about some other way to save this kind of cached data for ISL.

Also, this would be better with "frecency" instead of just frequency as the sort key. That way you could shift your reviewers over time and it would update without taking a long time. I'll do this later.

Reviewed By: quark-zju

Differential Revision: D47743089

fbshipit-source-id: b253d717d5195d0c521da95ffb2cb2beb9baad01