Cache suggested reviewers
Summary:
Since we  based our suggested reviewers on a selector, which used uncommitted changes, it would refetch very frequently. While this could ostensibly change the set of suggested reviewers, it's a bit over the top in making the UI flicker with a spinner as you do normal edits. We should instead cache the list and only recheck periodically.

To do this I added a cache layer in the selector. I don't know that there's a better way to do this in recoil. It's slightly silly becuase we're duplicating our data which the selector itself also caches. But it's a tiny amount of data so I don't think it matters.

Reviewed By: quark-zju

Differential Revision: D47743085

fbshipit-source-id: 3cebd3639352ed5ba0872302d544b61b3e73a9c1