Add component for suggested reviewers
Summary:
Add a component to render suggested reviewers and recent reviewers. The idea is that we render one short row of suggested/recent, and you can click to add them. We fit as many in as possible without wrapping (this is done with CSS)

This diff just uses some fake data for the reviewers.

Reviewed By: quark-zju

Differential Revision: D47743088

fbshipit-source-id: 0c39a5389bdbf0484b4344eac3a1c450f58a6703