Hook up server fetching for suggested reviewers
Summary:
Scaffolding to send a message to the server to do the suggested reviewers, then send an answer back, and show a loading spinner while waiting, and also use the code review provider to do the suggestion fetch, and otherwise skip it if it's not supported.

Real fetch in the next diff.

Reviewed By: quark-zju

Differential Revision: D47743086

fbshipit-source-id: aff51302f08fcdc575a68f0548686b47f64e2ab7