hook: support "background:" to run hook in background
Summary:
Add `background:` as a special prefix for hooks to kick off something in
background. This is intended to make certain edenfs-update hooks run in
background to improve `update` performance esp on Windows.

Note: while `something &; disown` can work for bash to run something in
background, it is annoyingly not cross-platform. The `background:` provides
a cross-platform way to run something.

Reviewed By: sggutier

Differential Revision: D47804790

fbshipit-source-id: 0ec5da4f6e56efd8388aedfa6d1204198579dfde