rebase: clarify that commits that become empty are skipped
Summary:
Our message just said that it created no changes, but we didn't
explicitly say that that meant that the the commit was not rebased. It
also wasn't clear why it created no changes, so this diff also
clarifies that that was because the changes were already in the
destination.

The new message was updated based on the [upstream
patch](https://repo.mercurial-scm.org/hg/diff/4edd427f34c1/hgext/rebase.py)

Reviewed By: evangrayk

Differential Revision: D47812328

fbshipit-source-id: e0be4e740eabf75b01810ba6ef68f7c94e07263f