ZLib encoding of Git object data for Packfile generation
Summary: This diff adds helper method in `PackfileItem` type that writes Zlib encoded object data to the provided buffer. This will be useful when appending raw object content to the packfile. The next diff in the stack will add a helper method that writes the object header as per the packfile format.

Differential Revision: D47837100

fbshipit-source-id: 20e7f8a4fa8e7637da0d1d4351f5425fa6e1c8fb