symlinks: add additional gating for TreeEntry:: getDType
Summary:
Some tools seem to query EdenFS directly about the type of certain files. If queried directly, EdenFS might answer that some files are symlinks regardless of whether symlink support for Windows is enabled or not. This adds some additional gating to make sure that EdenFS only answers that files are symlinks if symlink support is enabled for a checkout.

Split between this diff and D47829752 for clarity

Reviewed By: xavierd

Differential Revision: D47326227

fbshipit-source-id: 6e60b3d434ba45dd529286436526350ee0ac9dcf