symlinks: add additional gating for TreeEntry::getType
Summary:
Some tools seem to query EdenFS directly about the type of certain files. If queried directly, EdenFS might answer that some files are symlinks regardless of whether symlink support for Windows is enabled or not. This adds some additional gating to make sure that EdenFS only answers that files are symlinks if symlink support is enabled for a checkout.

Split between this diff and D47326227 for clarity

Reviewed By: xavierd

Differential Revision: D47829752

fbshipit-source-id: e66c7d774b5fd037fd26538f945e2ff077f3975d