revset: remove followlines revset
Summary:
It is not used in the past month. It is the only user of the revlog descendants.
Remove it to further clean up revlog.

This does not affect `log --line-range` though, which uses a different code path.

In the future we might want to use linelog for line-range queries.

Reviewed By: muirdm

Differential Revision: D47875335

fbshipit-source-id: 6420cca0718db80b83b0fd65ca21e615a243c6ef