Invalidate generated message cache if head hash changes
Summary:
We're caching generated commit messages by commit hash, or, if you're writing a new commit message, by the string "head". Caching like this would mean you could make a new commit, then start a new commit, and if you opened the generate dialog again it would still have the previous message, which may no longer apply.

So we keep track of the acutal head commit if you're in commit mode, and refuse to use the cache if the hash changes.

Reviewed By: quark-zju

Differential Revision: D47778029

fbshipit-source-id: 5661d29b01e27ca3d5f214e4d9848ab04bcfc4ba