Simplify generated message caching
Summary:
Previously, we had some complicated logic to cache the generated commit message when writing a new commit message. It involved looking up the head commit hash and using that to explicitly invalidate the cache when the head commit changes. This can be simplified by making the recoil cache key based on the hash instead, since we already have access to the base hash when we create the key to lookup the recoil state.

While doing this, I also simplified some other data being sent to the server by sending a comparison instead, and removing unnecessary data.

Reviewed By: quark-zju

Differential Revision: D47845157

fbshipit-source-id: 6238a97d4480da8636c6077500b6ed2a17e4a4b8