Unlink tools, leverage new unlink implementation on Multiplex
Summary:
Now that we can unlink from a multiplexed blobstore by using the `BlobstoreUnlinkOps` trait, we need
substantially less custom code in `mononoke newadmin blobstore-unlink` and `mononoke admin blobstore-bulk-unlink`.

Refactor this code to take advantage of this fact.

The code that was deleted is mostly duplicated with code that's present in `blobstore_factory` and
exposed if various ways (`repo_factory`, a facet, now `mononoke_app`).

There is still some custom code in `blobstore-unlink` as we remove the wrapper blob configs so we can see the keys
as `repo0000.key` instead of just `key`.
We also kept the historical ability of selecting only one underlying blobstore at a time.

For that reason, we expose a way to override the blob_config from a repo in `MononokeApp`. I'd have
preferred not to do that and simply using the facet, but it's the price of backward compatibility.

Differential Revision: D47799730

fbshipit-source-id: b1ebbf8bde4a98d9500dd4c3dd3fa8f3e54ca387