Add debug tool to console.log server/client messages
Summary:
It's very useful to see the list of messages sent between client and server. Add a checkbox to the internal debug tools to start logging messages.

Currently this is default-off. I could imagine persisting the setting to localstorage so that you could turn it on and refresh the page to see startup messages.

Reviewed By: quark-zju

Differential Revision: D47881299

fbshipit-source-id: 7d8dbca6d362f14285f16bdf55d769810d302358