Add type for operation progress
Summary:
This is the start of a stack to add more granular progress information from some commands that support it. This includes progress bars and inline progress information for a given commit hash.

Here I start by adding a new way for operations to return progress, just like it would stdout or stderr. This new channel will behave like a special form of stdout which is capable of including a progress bar.

Reviewed By: quark-zju

Differential Revision: D47888586

fbshipit-source-id: 5a574678bccfc2303a5c0a378283bd1fca7f403d