Internal debugging tools
Summary: Add infra to show a new dropdown menu for debugging tools. The button to show this menu will be hidden by default in prod, but visible by default in dev mode. There will be a setting to enable it even in prod mode. This isn't particularly expensive since we can lazy require the entire debug dropdown contents, which means it's code split into a new bundle.

Reviewed By: quark-zju

Differential Revision: D47881302

fbshipit-source-id: e53be95d8c15e03e83bc9aa97cbf627a0279bc74