Take sharded-scope-name argument
Summary:
We need for the backsyncer to take the `sharded-scope-name` CLI argument in so we can specify the regions it runs on, see D47917826 (currently hard-coded to global).

It currently uses the old `cmdlib` framework.

The long-term fix would be to migrate it to the new clap syntax and re-use [`ShardedExecutorArgs`](https://www.internalfb.com/code/fbsource/[0f4e49d078c898f56969344e50bf3b3fc87aa2ed]/fbcode/eden/mononoke/cmdlib/sharding/src/args.rs?lines=28)

To minimize the risk of backward-compatibility breakage as we want this to be set-up for the www source of truth change early next week, bring the legacy framework to parity with the new framework with respect to this command line argument.

The new argument is not required, so we don't break users of the old framework.

Reviewed By: RajivTS

Differential Revision: D47920862

fbshipit-source-id: de8a72f84266afc43148750e6da7748f3d22686d