Update gitoxide crates from git- to gix-
Summary: The family of `gitoxide` crates have moved from `git-` prefix to `gix-` prefix and the `git-` crates will no longer receive any updates. This diff removes the older crates and uses the newer `gix-` ones with the latest current version. As part of the migration, I had to include a new `gix-date` crate for the `Time` type which was earlier part of `gix-actor`.

Reviewed By: zertosh

Differential Revision: D47874915

fbshipit-source-id: 41ebb5396d8e6b9af6afadcb6b3ecb7fd3cc740e