bundlerepo: further improve test compatibility
Summary:
The bundlerevlog is tricky. It inserts raw chunk offsets to the "revlog index"
as a way to make the "revlog" away of the "graph" change, and use the indexes
to figure out where to seek and read. However, the main problem of this
approach is that if there are multiple "bundlerevlog" that uses the same
"revlog" under the hood, the added parts in the index might mess up one
bundlerevlog or the other. In some tests (ex. test-fb-ext-pushrebase.t),
this could cause problems like "revision not found".

Let's track the "node -> rawtext" mappings explicitly and use that to answer
"revision" to skip the mess up part of indexes. This increases memory usage but
our bundle usage in production should be already lazy (no files or trees) and
commits are already expanded in a similar way (see the handling of "inner"), so
it's likely okay.

Reviewed By: zzl0

Differential Revision: D47935352

fbshipit-source-id: 7f00b8b66c836082843a062c39722bd0b9fb12e2