Show tooltip explaining why drag and drop is disabled
Summary:
We don't allow drag & drop rebase with uncommitted changes, since it can cause conflicts. But we wouldn't tell the user this, so it was hard to know what was happening.

So, let's show a brief tooltip explaining that you can't rebase with uncommitted changes.

In the old ISL we had this at the bottom of the screen but it was pretty hard to see, this is much more contextual.

1.5s seems about right to me as long enough to read the message but not too long to get in the way.

also, we can allow drag rebasing if you have only untracked changes, so ignore those in the computation.

Reviewed By: quark-zju

Differential Revision: D47959148

fbshipit-source-id: 153902f172b6f84177b50c439b98327b3f6dda16