[isl] Add badge for GitHub PR review decision
Summary: This diff relates to [Issue #652](https://github.com/facebook/sapling/issues/652).

It's useful to know the review state of a PR. We do this for diffs internally, and in fact that's the main "status" we show. GitHub differentiates between "Open/Closed/Merged" but separately differentiates independent review states "Approved/Changes Requested/Review Required".

We should show this review state as well as status.

I considered merging the two, so it's more like we have internally. But I think it makes more sense to keep this separate, to be true to GitHub's data model.

Test Plan: {F1062573176} {F1062573175}

Reviewers: #isl,

Subscribers:

Tasks:

Tags:

Differential Revision: undefined