symlinks: make EdenFS able to deal with directory symlinks
Summary:
This diff makes it possible to check out commits with directory symlinks on Windows. Previously checking out commits with symlinks would make these symlinks appear as symlinks to regular files instead of symlinks to directories, making them unusable.

In order to solve this, each time a new symlink is to be created by ProjFS, we check the type of the directory with the equivalent of the [`stat()` function](https://pubs.opengroup.org/onlinepubs/000095399/functions/stat.html). This recursively reads symlinks until the target directory is a symlink or a regular file. Note that this populates all symlinks in the chain and materializes the final file / directory.

For avoiding cycles, paths are marked as "visited", so that if a cycle is found, we mark that entire cycle as a cycle of symlinks pointing to regular files. This was decided arbitrarily; the type of the cycle doesn't matter as symlinks cycles are useless anyways.

Reviewed By: xavierd

Differential Revision: D47144225

fbshipit-source-id: d521ab0d561cfe452675ae7249476879a1d3bedf