tests: make commitcloud tests more compatible with windows
Summary: On Windows `test-commitcloud-background-logging-perms.t` was failing when running it with buck, as it was trying to use the `hg` command, which is not available on `$TESTTMP/bin` unlike on POSIX due to how `debugruntest` works. This diff changes it so that commitcloud uses `util.hgcmd()` instead of the `hg` command directly.

Reviewed By: quark-zju

Differential Revision: D47839723

fbshipit-source-id: 0810a01dbfd93f7cd7c13fed14c0848e8bdbafbc