tests: make test-basic.t work on Windows
Summary:
On `test-basic.t` we use a regex for checking the output of a certain command. However, on Windows for debugruntest tests, we replace backward slashes with forward slashes for compatibility reasons.

This diff changes that, so that that replacement is only done when the line being checked is not a regex.

Reviewed By: quark-zju

Differential Revision: D47905816

fbshipit-source-id: fa93260b41a8b4ece53eefe1dc1beb93d329bef5