tests: make test-fb-ext-smartlog.t run under Buck on Windows
Summary: Due to the .bat hackery we have around hg when running it on Buck on Windows, newline characters are sometimes not properly passed down to hg commands when using `inprocess-hg-incompatible` on `debugruntest`. `test-fb-ext-smartlog.t` is one of these tests, and in one line it uses a template for separating output in different lines. We can enable this test by changing newlines by spaces, which should keep the functionality of that test.

Reviewed By: quark-zju

Differential Revision: D47905803

fbshipit-source-id: 9c3e9c9842ae2df16e741bdb4d174a8710ee0dfb