Remove SQL backed redaction store
Summary:
The tunable for it (`redaction_config_from_xdb`) has been disabled for a long time.
We know the blobstore backed implementation works well since it's been in production for a while.
Let's remove some old cruft.

Because the SQL backend was used in unit tests due to ease of setup, I added an implementation backed by a HashMap to keep this property.

Also delete the old mononoke_admin command.
It was already obsolete since `newadmin redaction` exists and is backed by the modern implementation.

Reviewed By: RajivTS

Differential Revision: D47872270

fbshipit-source-id: d7dbc1975208521a4fe0f9d8f87f185bd814ac1d