Remove unintentional conflict between `--sharded-scope-name` and `--sharded-service-name`
Summary:
`--sharded-scope-name` was added recently (D47920862) to allow the old CLI framework to be compatible with running on shards in specific regions.

It took a bit of time to properly test the change as we needed to deploy it through conveyor, then setup a shard manager config and a tw config.

The verdict is out: it is buggy. (And caused the [new tw tasks](https://www.internalfb.com/tupperware/job/summary?handle=tsp_frc%2Fmononoke%2Fmononoke_x_repo_backsync&start=1690978240&end=1690980268) to crash-loop until they were `tw stop`ped).

We unintentionally configured it so that `--sharded-scope-name` and `--sharded-service-name` conflict by making them members of the same clap group.
These flags are meant to be used together.

Fix this bug.

Reviewed By: liubov-dmitrieva

Differential Revision: D47993735

fbshipit-source-id: 50ae442207d18057c1e6a2ad1b365cde9943da73