Fix crash when dumping conflict info
Summary:
I saw a report of someone hitting issues with merge conflcits in ISL (both new and old). We were trying to decode binary files as utf8, which they are not guaranteed to be. We don't really need the file contents if it's a binary file. So let's just put None instead for the contents.

We actually had tests for binary files, but they just used the null byte, which is valid utf-8. We can instead insert some invalid unicode. The simplest I found is `0xE1` at the end of the file, which is an unfinished 2-byte sequence making it all invalid utf8. This way our tests exercise this code path.

Reviewed By: quark-zju

Differential Revision: D48008610

fbshipit-source-id: 9aa42db0991bad456c2fb9bd998fbc0597bea5c7